+++ contracts/Logo.sol
@@ -5,38 +5,31 @@ import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 import "@openzeppelin/contracts/token/ERC721/extensions/ERC721Enumerable.sol";
 import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
-import "@openzeppelin/contracts/utils/Counters.sol";
 
 /// @custom:security-contact security@dlogos.xyz
 contract Logo is ERC721, ERC721Enumerable, ERC721URIStorage, Ownable {
-    using Counters for Counters.Counter;
+    uint256 public _tokenIdCounter; // Starting from 1
 
-    Counters.Counter private _tokenIdCounter;
-
-    constructor() ERC721("Logo", "LOGO") {}
+    constructor() ERC721("Logo", "LOGO") Ownable(msg.sender) {}
 
     function _baseURI() internal pure override returns (string memory) {
         return "https://gateway.pinata.cloud/ipfs/";
     }
 
     function safeMint(address to, string memory uri) public onlyOwner {
-        uint256 tokenId = _tokenIdCounter.current();
-        _tokenIdCounter.increment();
-        _safeMint(to, tokenId);
-        _setTokenURI(tokenId, uri);
+        _tokenIdCounter++;
+        _safeMint(to, _tokenIdCounter);
+        _setTokenURI(_tokenIdCounter, uri);
     }
 
     // The follong functions are overrides required by Solidity.
 
-    function _beforeTokenTransfer(address from, address to, uint256 tokenId, uint256 batchSize)
-        internal
-        override(ERC721, ERC721Enumerable)
-    {
-        super._beforeTokenTransfer(from, to, tokenId, batchSize);
+    function _update(address to, uint256 tokenId, address auth) internal override(ERC721, ERC721Enumerable) returns (address) {
+        ERC721Enumerable._update(to, tokenId, auth);
     }
 
-    function _burn(uint256 tokenId) internal override(ERC721, ERC721URIStorage) {
-        super._burn(tokenId);
+    function _increaseBalance(address account, uint128 amount) internal override(ERC721, ERC721Enumerable) {
+        ERC721Enumerable._increaseBalance(account, amount);
     }
 
     function tokenURI(uint256 tokenId)